Updated pipette functions and Jubilee controller functions. Removed old files from name change
