+++ science_jubilee/decks/deck_definition/lab_automation_deck.json
@@ -1,0 +1,64 @@
+{
+    "bedType": "Lab Automation Plate",
+    "deckSlots": {
+        "total": 6,
+        "type": "SLAS Standard Labware"
+    },
+    "slots": {
+        "0": {
+            "offset": [
+                13.8,
+                5
+            ],
+            "has_labware": false,
+            "labware": null
+        },
+        "1": {
+            "offset": [
+                153.5,
+                5.4
+            ],
+            "has_labware": false,
+            "labware": null
+        },
+        "2": {
+            "offset": [
+                14,
+                103.3
+            ],
+            "has_labware": false,
+            "labware": null
+        },
+        "3": {
+            "offset": [
+                153.8,
+                102.4
+            ],
+            "has_labware": false,
+            "labware": null
+        },
+        "4": {
+            "offset": [
+                14.3,
+                199.6
+            ],
+            "has_labware": false,
+            "labware": null
+        },
+        "5": {
+            "offset": [
+                154.2,
+                199.7
+            ],
+            "has_labware": false,
+            "labware": null
+        }
+    },
+    "offsetFrom": {
+        "corner": "Bottom Left"
+    },
+    "material": {
+        "plate": "Aluminum",
+        "mask": "Delrin"
+    }
+}
