+++ science_jubilee/labware/labware_definition/20mlscintillation_12_wellplate_18000ul.json
